Squashing support of 2.7 for firefox, Removed firefox_webdriver.py
